Added time table, step time enum, color enum, changes to pack/unpack functions. Added commands class. NOTE: pack/unpack was not used before, and is broken currently.

+++ DmxLightShow/MestraTypes.h
@@ -26,7 +26,7 @@ typedef uint8_t  dmx_value_t;
 typedef uint8_t  program_t;
 typedef uint8_t  preset_t;
 typedef  int16_t step_t; // Can be negative (for rainbow program)
-typedef uint16_t step_duration_t;
+typedef uint8_t  step_duration_t;
 typedef uint32_t step_time_t;
 typedef  int16_t parameter_t;
 typedef uint8_t  intensity_t;
